Watch props to create music controls
It should now create music controls when the title and channel name
are available, and then, again when the thumbnail loads. The reason
these are separate calls is because I ran into issues in Android 11. Sometimes,
the thumbnail loading seems to prevent the music controls from showing.
So, I call it twice, and as a result, need to destroy it twice.